+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Hide add icon on decks and models section.
+- Drawer not closing on route change on mobile.
 
 ## [0.3.2] - 2019-05-06
 ### Changed